refactor : separate packages in model
