+++ src/main/java/lotto/model/prize/PrizeJudge.java
@@ -1,0 +1,12 @@
+package lotto.model.prize;
+
+public class PrizeJudge {
+    public Prize getPrizeOf(int matchCount, boolean hasBonus) {
+        if (matchCount == 6) return Prize.FIRST;
+        if (matchCount == 5 && hasBonus) return Prize.SECOND;
+        if (matchCount == 5) return Prize.THIRD;
+        if (matchCount == 4) return Prize.FOURTH;
+        if (matchCount == 3) return Prize.FIFTH;
+        return Prize.LOSING;
+    }
+}
