+++ src/main/java/lotto/model/prize/PrizeRecord.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.prize;
 
 import java.util.HashMap;
