+++ src/main/java/lotto/model/ticket/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.ticket;
 
 import java.util.HashMap;
 
@@ -6,7 +6,7 @@ public class LottoNumbers {
     private final HashMap<Integer, LottoNumber> lottoNumbers;
 
     public LottoNumbers() {
-        this.lottoNumbers = new HashMap();
+        this.lottoNumbers = new HashMap<>();
         for (int i = 1; i <= LottoNumber.NUMBER_RANGE; i++) {
             this.lottoNumbers.put(i, new LottoNumber(i));
         }
