+++ src/main/java/lotto/model/ticket/LottoTicket.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.ticket;
 
 import java.util.ArrayList;
 import java.util.HashSet;
