+++ src/test/java/lotto/model/ticket/LottoNumbersTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.ticket;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
